+++ src/app/managers/mutations.ts
@@ -1,25 +1,26 @@
-import { ICuiMutationHandler } from "../../core/models/interfaces";
-import { is, getFirstAttributeStartingWith } from "../../core/utlis/functions";
+import { ICuiMutationHandler, IUIInteractionProvider } from "../../core/models/interfaces";
+import { is, getFirstAttributeStartingWith, getMutationAttribute } from "../../core/utlis/functions";
 import { ATTRIBUTES } from "../../core/utlis/statics";
 import { CuiIconHandler } from "../handlers/icon";
 import { CuiSpinnerHandler } from "../handlers/spinner";
+import { CuiCircleHandler } from "../handlers/circle";
 
 export class CuiAttributeMutationHandler {
-    public static get(element: any): ICuiMutationHandler {
+    public static get(element: any, interactions?: IUIInteractionProvider): ICuiMutationHandler {
         if (!is(element)) {
             return null;
         }
-        let attribute = getFirstAttributeStartingWith(element, 'data')
+        let attribute = getMutationAttribute(element)?.name
         if (!is(attribute)) {
             return null;
         }
-        switch (attribute.name) {
+        switch (attribute) {
             case ATTRIBUTES.icon:
-                return new CuiIconHandler(element);
-            case ATTRIBUTES.scale:
-                return new CuiIconHandler(element);
+                return new CuiIconHandler(element, interactions);
             case ATTRIBUTES.spinner:
-                return new CuiSpinnerHandler(element);
+                return new CuiSpinnerHandler(element, interactions);
+            case ATTRIBUTES.circle:
+                return new CuiCircleHandler(element, interactions);
         }
         return null;
     }