+++ changelog.md
@@ -1,5 +1,12 @@
 # bpd cUI Light changelog
-
+## [0.0.4] 2020-06-11
+* [Removed] Grid system from grid display mode
+* [Fixed] Dropdown display
+* [Fixed] Svg not showing in dark mode
+* [Changed] Bahavior of mutation handler
+* [Added] Interaction handler to mutated elements
+* [Added] Collection handler to handle elements collections like lists, tabs switchers
+* [Added] Circle progress
 ## [0.0.3] 2020-06-10
 * [Fixed] Force element to be a flex when *-flex
 * [Changed] Behavior of width / visiblity / container elements